+++ osp/energymon-osp.h
@@ -10,18 +10,19 @@
 extern "C" {
 #endif
 
+#include <inttypes.h>
 #include <stddef.h>
 #include "energymon.h"
 
 int energymon_init_osp(energymon* em);
 
-unsigned long long energymon_read_total_osp(const energymon* em);
+uint64_t energymon_read_total_osp(const energymon* em);
 
 int energymon_finish_osp(energymon* em);
 
 char* energymon_get_source_osp(char* buffer, size_t n);
 
-unsigned long long energymon_get_interval_osp(const energymon* em);
+uint64_t energymon_get_interval_osp(const energymon* em);
 
 int energymon_get_osp(energymon* impl);
